// js_modules/dagit/packages/core/src/launchpad/types/ConfigEditorGeneratorPipelineFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: ConfigEditorGeneratorPipelineFragment
// ====================================================

export interface ConfigEditorGeneratorPipelineFragment_presets_tags {
  __typename: "PipelineTag";
  key: string;
  value: string;
}

export interface ConfigEditorGeneratorPipelineFragment_presets {
  __typename: "PipelinePreset";
  name: string;
  mode: string;
  solidSelection: string[] | null;
  runConfig: any;
  tags: ConfigEditorGeneratorPipelineFragment_presets_tags[];
}

export interface ConfigEditorGeneratorPipelineFragment_tags {
  __typename: "PipelineTag";
  key: string;
  value: string;
}

export interface ConfigEditorGeneratorPipelineFragment {
  __typename: "Pipeline";
  id: string;
  isJob: boolean;
  name: string;
  presets: ConfigEditorGeneratorPipelineFragment_presets[];
  tags: ConfigEditorGeneratorPipelineFragment_tags[];
}